+++ qcengine/programs/tests/test_entos.py
@@ -1,8 +1,8 @@
 import pytest
-
 import qcelemental as qcel
-import qcengine as qcng
 from qcelemental.testing import compare_recursive
+
+import qcengine as qcng
 from qcengine.testing import qcengine_records, using_entos
 
 entos_info = qcengine_records('entos')
@@ -13,12 +13,12 @@ def test_entos_output_parser(test_case):
 
     # Get output file data
     data = entos_info.get_test_data(test_case)
-    inp = qcel.models.ResultInput.parse_raw(data["input.json"])
+    inp = qcel.models.AtomicInput.parse_raw(data["input.json"])
 
     output = qcng.get_program('entos', check=False).parse_output(data, inp).dict()
     output.pop("provenance", None)
 
-    output_ref = qcel.models.Result.parse_raw(data["output.json"]).dict()
+    output_ref = qcel.models.AtomicResult.parse_raw(data["output.json"]).dict()
     output_ref.pop("provenance", None)
 
     check = compare_recursive(output_ref, output)
@@ -30,7 +30,7 @@ def test_entos_input_formatter(test_case):
 
     # Get input file data
     data = entos_info.get_test_data(test_case)
-    inp = qcel.models.ResultInput.parse_raw(data["input.json"])
+    inp = qcel.models.AtomicInput.parse_raw(data["input.json"])
 
     # TODO add actual comparison of generated input file
     input_file = qcng.get_program('entos', check=False).build_input(inp, qcng.get_config())
@@ -42,7 +42,7 @@ def test_entos_input_formatter_template(test_case):
 
     # Get input file data
     data = entos_info.get_test_data(test_case)
-    inp = qcel.models.ResultInput.parse_raw(data["input.json"])
+    inp = qcel.models.AtomicInput.parse_raw(data["input.json"])
 
     # TODO add actual comparison of generated input file
     input_file = qcng.get_program('entos', check=False).build_input(inp, qcng.get_config(), template="Test template")
@@ -54,14 +54,14 @@ def test_entos_input_formatter_template(test_case):
 def test_entos_executor(test_case):
     # Get input file data
     data = entos_info.get_test_data(test_case)
-    inp = qcel.models.ResultInput.parse_raw(data["input.json"])
+    inp = qcel.models.AtomicInput.parse_raw(data["input.json"])
 
     # Run entos
     result = qcng.compute(inp, 'entos')
     assert result.success is True
 
     # Get output file data
-    output_ref = qcel.models.Result.parse_raw(data["output.json"])
+    output_ref = qcel.models.AtomicResult.parse_raw(data["output.json"])
 
     atol = 1e-6
     assert compare_recursive(output_ref.return_result, result.return_result, atol=atol)