+++ qcengine/programs/tests/test_molpro.py
@@ -2,10 +2,10 @@ import copy
 
 import numpy as np
 import pytest
-
 import qcelemental as qcel
-import qcengine as qcng
 from qcelemental.testing import compare_recursive
+
+import qcengine as qcng
 from qcengine.testing import qcengine_records, using_molpro
 
 molpro_info = qcengine_records('molpro')
@@ -16,12 +16,12 @@ def test_molpro_output_parser(test_case):
 
     # Get output file data
     data = molpro_info.get_test_data(test_case)
-    inp = qcel.models.ResultInput.parse_raw(data["input.json"])
+    inp = qcel.models.AtomicInput.parse_raw(data["input.json"])
 
     output = qcng.get_program('molpro', check=False).parse_output(data, inp).dict()
     output.pop("provenance", None)
 
-    output_ref = qcel.models.Result.parse_raw(data["output.json"]).dict()
+    output_ref = qcel.models.AtomicResult.parse_raw(data["output.json"]).dict()
     output_ref.pop("provenance", None)
 
     # TODO add `skip` to compare_recursive
@@ -34,7 +34,7 @@ def test_molpro_input_formatter(test_case):
 
     # Get input file data
     data = molpro_info.get_test_data(test_case)
-    inp = qcel.models.ResultInput.parse_raw(data["input.json"])
+    inp = qcel.models.AtomicInput.parse_raw(data["input.json"])
 
     # TODO add actual comparison of generated input file
     input_file = qcng.get_program('molpro', check=False).build_input(inp, qcng.get_config())
@@ -46,14 +46,14 @@ def test_molpro_input_formatter(test_case):
 def test_molpro_executor(test_case):
     # Get input file data
     data = molpro_info.get_test_data(test_case)
-    inp = qcel.models.ResultInput.parse_raw(data["input.json"])
+    inp = qcel.models.AtomicInput.parse_raw(data["input.json"])
 
     # Run Molpro
     result = qcng.compute(inp, 'molpro', local_options={"ncores": 4})
     assert result.success is True
 
     # Get output file data
-    output_ref = qcel.models.Result.parse_raw(data["output.json"])
+    output_ref = qcel.models.AtomicResult.parse_raw(data["output.json"])
 
     atol = 1e-6
     assert compare_recursive(output_ref.return_result, result.return_result, atol=atol)